squared relu works for random search, but not backpropogation
i would like to get back prop to work... but the more I think about it,
the more I don't think it is necessary. Relu squared is functional and
the tests that I have done on AND, XOR, and XNOR have converged to a
proper solution. All backprop tests fail, presumably because I don't
know what I am doing. However, I want to move to a 1-bit architecture,
which removes the weights entirely... so that means that some delta
value to propogate back on the weights will go away. so does that mean
back prop goes away? probably not. there is probably some nice way to
handle converting a back prop delta into a add/subtract condition; but I
don't know what that is, and I think I am going to push that onto future
me.

Which means. the next several commits are going to be some serious
surgery.